chore: split up model classes into different files.
