+++ conftool/model/date_field.py
@@ -1,0 +1,5 @@
+from .config_enum import ConfigEnum
+
+class DateFieldModel(ConfigEnum):
+    LOGDATE = "LogDate"
+    CREATEDDATE = "CreatedDate"
