+++ conftool/model/env.py
@@ -1,0 +1,9 @@
+from .base import BaseModel
+
+class EnvModel(BaseModel):
+    end_date: str
+    start_date: str
+
+    def check(self):
+        # No checks required
+        return super().check()
