+++ conftool/model/generation_interval.py
@@ -1,0 +1,5 @@
+from .config_enum import ConfigEnum
+
+class GenerationIntervalModel(ConfigEnum):
+    HOURLY = 'Hourly'
+    DAILY = 'Daily'
