+++ conftool/model/grant_type.py
@@ -1,0 +1,5 @@
+from .config_enum import ConfigEnum
+
+class GrantTypeModel(ConfigEnum):
+    PASSWORD = "password"
+    JWT = "urn:ietf:params:oauth:grant-type:jwt-bearer"
