+++ conftool/model/query.py
@@ -1,21 +1,14 @@
 from .base import BaseModel
-from .api_ver import ApiVer
+from .api_ver import ApiVerModel
 from .exception import ConfigException
-
-class Env(BaseModel):
-    end_date: str
-    start_date: str
-
-    def check(self):
-        # No checks required
-        return super().check()
+from .env import EnvModel
 
 class QueryModel(BaseModel):
     query: str
     timestamp_attr: str
     rename_timestamp: str
-    api_ver: ApiVer
-    env: Env
+    api_ver: ApiVerModel
+    env: EnvModel
     api_name: str
     event_type: str
     id: list[str]
